Move Base.show/summary methods into a single show_methods.jl file
Consolidates the Base.show and Base.summary methods added across
Domains, Parameters, boundary_conditions, models, drivers, LandModel,
Bucket, Snow, Soil, Runoff, and Canopy into one file, included last in
ClimaLand.jl so every referenced type/submodule is already defined.


Move show-related helper functions into show_methods.jl

_domain_device_str, _domain_show_fieldnames, _scalar_or_field_str,
_flatten_symbols, _kind_str, and _driver_kind only exist to support the
Base.show methods, so they move alongside them.


remove leftover line changes
