+++ src/shared_utilities/drivers.jl
@@ -155,62 +155,6 @@ struct PrescribedAtmosphere{
     end
 end
 
-# The precipitation/temperature/wind/humidity/pressure/co2 fields hold
-# TimeVaryingInputs backed by NetCDF readers, whose default show output can
-# run into the millions of characters; print only the kind of driver
-# (e.g. data-backed vs. analytic) instead of the driver's own contents.
-_driver_kind(x) = nameof(typeof(x))
-
-function Base.show(io::IO, ::MIME"text/plain", atmos::PrescribedAtmosphere)
-    if get(io, :compact, false)
-        show(io, atmos)
-    else
-        println(
-            io,
-            nameof(typeof(atmos)),
-            "{",
-            typeof(atmos).parameters[1],
-            "}",
-        )
-        println(io, "  start_date: ", atmos.start_date)
-        println(
-            io,
-            "  reference height: ",
-            atmos.h,
-            " m, gustiness: ",
-            atmos.gustiness,
-            " m/s",
-        )
-        driver_fields = (:liquid_precip, :snow_precip, :T, :u, :q, :P)
-        kinds = unique(_driver_kind(getfield(atmos, f)) for f in driver_fields)
-        if length(kinds) == 1
-            println(
-                io,
-                "  drivers: ",
-                kinds[1],
-                " (liquid/snow precip, T, u, q, P)",
-            )
-        else
-            println(io, "  drivers: ", join(kinds, ", "))
-        end
-        println(io, "  CO2: ", _driver_kind(atmos.c_co2))
-    end
-end
-
-function Base.show(io::IO, atmos::PrescribedAtmosphere)
-    print(
-        io,
-        nameof(typeof(atmos)),
-        "{",
-        typeof(atmos).parameters[1],
-        "}(start_date=",
-        atmos.start_date,
-        ")",
-    )
-end
-
-Base.summary(io::IO, atmos::PrescribedAtmosphere) = show(io, atmos)
-
 """
     PrescribedPrecipitation{FT, LP} <: AbstractAtmosphericDrivers{FT}
 Container for holding prescribed precipitation driver
@@ -736,61 +680,6 @@ struct PrescribedRadiativeFluxes{
         return new{FT, typeof.(args)...}(args...)
     end
 end
-
-# SW_d, frac_diff, and LW_d are TimeVaryingInputs backed by NetCDF readers,
-# whose default show output can run into the millions of characters; only the
-# scalar/metadata fields are cheap and informative to print.
-function Base.show(
-    io::IO,
-    ::MIME"text/plain",
-    radiation::PrescribedRadiativeFluxes,
-)
-    if get(io, :compact, false)
-        show(io, radiation)
-    else
-        println(
-            io,
-            nameof(typeof(radiation)),
-            "{",
-            typeof(radiation).parameters[1],
-            "}",
-        )
-        println(io, "  start_date: ", radiation.start_date)
-        sw_kind = _driver_kind(radiation.SW_d)
-        lw_kind = _driver_kind(radiation.LW_d)
-        if sw_kind == lw_kind
-            println(io, "  drivers (SW_d, LW_d): ", sw_kind)
-        else
-            println(io, "  drivers: SW_d=", sw_kind, ", LW_d=", lw_kind)
-        end
-        println(
-            io,
-            "  cosθs: ",
-            isnothing(radiation.cosθs) ? "computed from date and location" :
-            "prescribed",
-        )
-        println(
-            io,
-            "  frac_diff: ",
-            isnothing(radiation.frac_diff) ? "computed empirically" :
-            string("prescribed (", _driver_kind(radiation.frac_diff), ")"),
-        )
-    end
-end
-
-function Base.show(io::IO, radiation::PrescribedRadiativeFluxes)
-    print(
-        io,
-        nameof(typeof(radiation)),
-        "{",
-        typeof(radiation).parameters[1],
-        "}(start_date=",
-        radiation.start_date,
-        ")",
-    )
-end
-
-Base.summary(io::IO, radiation::PrescribedRadiativeFluxes) = show(io, radiation)
 
 """
     net_radiation!(dest::ClimaCore.Fields.Field,
