+++ src/shared_utilities/models.jl
@@ -56,60 +56,6 @@ as a `make_compute_imp_tendency` default.
 """
 abstract type AbstractExpModel{FT} <: AbstractModel{FT} end
 
-# prognostic_vars can be a flat tuple of symbols (e.g. soil) or a NamedTuple
-# grouping symbols by sub-component with many empty entries (e.g. canopy);
-# flatten to just the active variable names either way.
-_flatten_symbols(x::Symbol) = (x,)
-_flatten_symbols(x::Union{Tuple, NamedTuple}) =
-    isempty(x) ? () : Tuple(Iterators.flatten(map(_flatten_symbols, values(x))))
-
-# Many parameters (porosity, hydraulic conductivity, leaf reflectance, ...)
-# may be either a scalar or a non-scalar value (a ClimaCore Field varying in
-# space, or a TimeVaryingInput varying in time); non-scalar values have their
-# own default show that's far too long to embed inline in a parameter
-# struct's show method, so print a value for scalars and just the type name
-# otherwise.
-_scalar_or_field_str(x::AbstractFloat) = string(x)
-_scalar_or_field_str(x) = string(nameof(typeof(x)))
-
-# A closure/parameterization choice (e.g. a soil hydrology closure) may be
-# stored as a scalar struct, or as a Field of that struct when it varies in
-# space; either way, name the closure itself rather than saying "Field".
-_kind_str(x) = string(nameof(typeof(x)))
-_kind_str(x::ClimaCore.Fields.Field) = string(nameof(eltype(x)))
-
-# Generic fallback for any concrete model: models typically hold domains,
-# parameters, and boundary conditions (which can themselves hold prescribed
-# drivers), so a default field dump can be arbitrarily large. Concrete models
-# that want to show more than the type name and domain/prognostic variables
-# (e.g. composed models like LandModel) should add their own more specific
-# method.
-function Base.show(io::IO, ::MIME"text/plain", model::AbstractModel)
-    if get(io, :compact, false)
-        show(io, model)
-    else
-        println(
-            io,
-            nameof(typeof(model)),
-            "{",
-            typeof(model).parameters[1],
-            "}",
-        )
-        if hasfield(typeof(model), :domain)
-            println(io, "  domain: ", sprint(show, model.domain))
-        end
-        pv = _flatten_symbols(prognostic_vars(model))
-        if !isempty(pv)
-            println(io, "  prognostic variables: ", pv)
-        end
-    end
-end
-
-Base.show(io::IO, model::AbstractModel) =
-    print(io, nameof(typeof(model)), "{", typeof(model).parameters[1], "}")
-
-Base.summary(io::IO, model::AbstractModel) = show(io, model)
-
 """
     name(model::AbstractModel)
